[data-get][s]: data get r links from DataHub -refs #323 (#324)

+++ bin/data-get.js
@@ -66,30 +66,34 @@ const run = async () => {
       if(identifier.substr(-1) === '/' && identifier.length > 1) {
         identifier = identifier.slice(0, identifier.length - 1)
       }
-      // Try to guess owner and dataset name here. We're not loading Dataset object
-      // because we want to handle private datasets as well:
-      const idParts = identifier.split('/')
-      const owner = idParts[idParts.length - 2]
-      const name = idParts[idParts.length - 1]
-      const token = config.get('token')
-      pathToSave = path.join(owner, name)
+      // We assume that if /r/ is in identifier then it's r link.
+      if (identifier.includes('/r/')) {
+        pathToSave = await saveFileFromUrl(identifier, argv.format)
+      } else {
+        // Try to guess owner and dataset name here. We're not loading Dataset object
+        // because we want to handle private datasets as well:
+        const idParts = identifier.split('/')
+        const owner = idParts[idParts.length - 2]
+        const name = idParts[idParts.length - 1]
+        const token = config.get('token')
+        pathToSave = path.join(owner, name)
 
-      if (!checkDestIsEmpty(owner, name)) {
-        throw new Error(`${owner}/${name} is not empty!`)
+        if (!checkDestIsEmpty(owner, name)) {
+          throw new Error(`${owner}/${name} is not empty!`)
+        }
+        
+        /** For datasets from the datahub we get zipped version and unzip it.
+                - less traffic
+                - zipped version has a fancy file structure
+            #issue: https://github.com/datahq/datahub-qa/issues/86  */
+        const zipped_dataset_url  = `https://datahub.io/${owner}/${name}/r/${name}_zip.zip?jwt=${token}`
+        const archive_path = await saveFileFromUrl(zipped_dataset_url, 'zip')
+        // unzip archive into destination folder
+        fs.createReadStream(archive_path)
+          .pipe(unzip.Extract({ path: pathToSave }))
+          // removing the archive file once we extracted all the dataset files
+          .on('finish', () => {fs.unlinkSync(archive_path)})
       }
-
-      /** For datasets from the datahub we get zipped version and unzip it.
-              - less traffic
-              - zipped version has a fancy file structure
-          #issue: https://github.com/datahq/datahub-qa/issues/86  */
-      const zipped_dataset_url  = `https://datahub.io/${owner}/${name}/r/${name}_zip.zip?jwt=${token}`
-      const archive_path = await saveFileFromUrl(zipped_dataset_url, 'zip')
-      // unzip archive into destination folder
-      fs.createReadStream(archive_path)
-        .pipe(unzip.Extract({ path: pathToSave }))
-        // removing the archive file once we extracted all the dataset files
-        .on('finish', () => {fs.unlinkSync(archive_path)})
-
     } else { // If it is not a dataset - download the file
       if (parsedIdentifier.type === 'github' && !githubDataset) {
         identifier += `?raw=true`
